Precompile the flow folder
